+++ cfme/infrastructure/pxe/base.py
@@ -1,0 +1,7 @@
+import sentaku
+
+
+class PXEBase(object):
+    # ** This is literally jsut to describe the exists contextual method. We can do better here
+    # ** again WIP card raised high! This may go away.
+    exists = sentaku.ContextualMethod()
