+++ utils/appliance/__init__.py
@@ -12,6 +12,8 @@ from time import sleep
 from urlparse import ParseResult, urlparse
 from tempfile import NamedTemporaryFile
 
+import sentaku
+
 from cached_property import cached_property
 
 from werkzeug.local import LocalStack, LocalProxy
@@ -24,9 +26,11 @@ from utils.mgmt_system import RHEVMSystem
 from mgmtsystem.virtualcenter import VMWareSystem
 
 from cfme.common.vm import VM
+
 from fixtures import ui_coverage
 from fixtures.pytest_store import store
-from utils import api, conf, datafile, db, trackerbot, db_queries, ssh, ports
+
+from utils import api, conf, datafile, db, trackerbot, db_queries, ports, ssh as john_ssh
 from utils.datafile import load_data_file
 from utils.events import EventTool
 from utils.log import logger, create_sublogger, logger_wrap
@@ -46,9 +50,21 @@ if not RUNNING_UNDER_SPROUT:
     from utils.browser import browser_session
     from utils.hosts import setup_providers_hosts_credentials
 
+# ** We need our endpoints! Can this be done via some entrypoints so that they can become
+# ** pip modules to install....... who knows!? Does it make sense? I'm too tired to make
+# ** a judgement call on that :)
+from endpoints.ui import UIEndpoint
+from endpoints.db import DBEndpoint
+
 
 class ApplianceException(Exception):
     pass
+
+# ** Here we define our implementation names. It has been already noted that we should carefully
+# ** think about the wording so that implemented_for(ViaDB) reads either implemented_for(DB) or
+# ** implemented(ViaDB)
+ViaUI = sentaku.ImplementationName('ViaUI')
+ViaDB = sentaku.ImplementationName('ViaDB')
 
 
 class IPAppliance(object):
@@ -79,6 +95,34 @@ class IPAppliance(object):
         self.browser_steal = browser_steal
         self.container = container
         self._db_ssh_client = None
+
+        # ** browser is an endpoint object and we ***CURRENTLY*** (gosh I can't stress that
+        # ** enough) pass in the name of the attribute that holds the "session" so that we
+        # ** can access that in the endpoint via the "owner.<session>".
+        # ** Why is this needed?
+        # ** Iteration my friend iteration. We have not yet moved the sessions to be inside
+        # ** the endpoint objects. That IS on the roadmap, so these little calls will become
+        # ** simpler.
+        # ** Note also the complexity surrounding this that browser_session is a "new" session
+        # ** and that we still need to work out how to handle multiple browsers, for multiple
+        # ** appliances and how that interacts with Wharf etc.
+        self.browser = UIEndpoint('ui', 'browser_session', self)
+        self.db_session = DBEndpoint('db', 'db', self)
+        # ** this will be renamed <---- see I told you we had it covered.
+        self.sentaku_ctx = sentaku.ImplementationContext(
+            implementations=sentaku.AttributeBasedImplementations(self, {
+                ViaDB: 'db_session',
+                ViaUI: 'browser',
+            }),
+            default_choices=[
+                ViaDB,
+                ViaUI,
+            ],
+        )
+        # ** Above here, we simple "link" the endpoint implementation objects with their
+        # ** corresponding endpoint objects. If this seems a little confusing?? It is.
+        # ** The good thing is that a) we can probably make this better, and b) we will
+        # ** probably only need to do it in this file. W00t!
 
     def __repr__(self):
         return '{}({})'.format(type(self).__name__, repr(self.address))
@@ -398,7 +442,8 @@ class IPAppliance(object):
             'password': conf.credentials['ssh']['password'],
             'container': self.container,
         }
-        ssh_client = ssh.SSHClient(**connect_kwargs)
+        # ** Ack! you caught me.....
+        ssh_client = john_ssh.SSHClient(**connect_kwargs)
         # FIXME: propperly store ssh clients we made
         store.ssh_clients_to_close.append(ssh_client)
         return ssh_client
